feat(auth): add Zustand store for user tokens and refactor action function
