+++ apps/web/package.json
@@ -37,7 +37,8 @@
     "react-resizable-panels": "^3.0.6",
     "react-router": "^7.5.0",
     "sonner": "^2.0.7",
-    "tailwind-merge": "^3.3.1"
+    "tailwind-merge": "^3.3.1",
+    "zustand": "^5.0.8"
   },
   "devDependencies": {
     "@react-router/dev": "^7.5.0",
